added sending email function

+++ upload_staff_records.php
@@ -1,6 +1,14 @@
 <?php
 include("admin_header.php");
 
+use PHPMailer\PHPMailer\PHPMailer;
+use PHPMailer\PHPMailer\SMTP;
+use PHPMailer\PHPMailer\Exception;
+
+require_once __DIR__ . '/gmailsender/vendor/phpmailer/src/Exception.php';
+require_once __DIR__ . '/gmailsender/vendor/phpmailer/src/PHPMailer.php';
+require_once __DIR__ . '/gmailsender/vendor/phpmailer/src/SMTP.php';
+
 //-------------------------------upload csv------------------------------------------------------------------------------//
 if(isset($_POST["upload"]))
 {
@@ -9,20 +17,62 @@ if(isset($_POST["upload"]))
         $filename = explode(".", $_FILES['staff_file']['name']);//convert file name into array and store into $ file name variable
         if(end($filename) == "csv")//check if file is csv or not
         {//if csv
+            if(!empty($_POST['check_list'])) {
             //truncate all data from table
             $truncate = "TRUNCATE `tbl_staff_record`";
             mysqli_query($db, $truncate);
+            }
             //truncate all data from table
             $handle = fopen($_FILES['staff_file']['tmp_name'], "r");//file open function
             while($data = fgetcsv($handle))//file get csv function: fetch comma delimited data from csv and convert into array and store into $ variable
             {
+                $mail = new PHPMailer(true);
                 $staff_id = mysqli_real_escape_string($db, $data[0]); //data is cleaned by mysqli real escape function
                 $first_name = mysqli_real_escape_string($db, $data[1]);  
                 $last_name = mysqli_real_escape_string($db, $data[2]);
-               
-                $query="INSERT INTO tbl_staff_record (staff_id, first_name, last_name) 
-                VALUES ('$staff_id', '$first_name', '$last_name')";
-                        mysqli_query($db, $query);
+                $email = mysqli_real_escape_string($db, $data[3]);
+                $check1 = mysqli_query($db, "SELECT * FROM tbl_staff_record WHERE staff_id='{$staff_id}' AND first_name='{$first_name}' AND last_name='{$last_name}'");
+                if (mysqli_num_rows($check1)!=1){
+                $query="INSERT INTO tbl_staff_record (staff_id, first_name, last_name, email) 
+                VALUES ('$staff_id', '$first_name', '$last_name', '$email')";
+                if (mysqli_query($db, $query)){
+                    $check = mysqli_query($db, "SELECT * FROM tbl_staff_registry WHERE staff_id='{$staff_id}' AND first_name='{$first_name}' AND last_name='{$last_name}'");
+                    if (mysqli_num_rows($check)!=1){
+                        $fname = $first_name.' '.$last_name;
+                        $username = $staff_id;
+                        $password = rand(100000, 999999);
+                        $passwd = password_hash($password, PASSWORD_DEFAULT);
+                        $sql ="INSERT INTO tbl_staff_registry (staff_id, first_name, last_name, username, `password`, register_status) VALUES ('$staff_id', '$first_name', '$last_name', '$username', '$passwd', '0')";
+                        mysqli_query($db, $sql);
+
+                            $mail->isSMTP();
+                            $mail->Host = 'smtp.gmail.com';
+                            $mail->SMTPAuth = true;
+                            $mail->SMTPSecure = PHPMailer::ENCRYPTION_STARTTLS;
+                            $mail->Port = 587;
+                            
+                            $mail->Username = 'goldenstatecollege00@gmail.com'; // YOUR gmail email
+                            $mail->Password = 'goldenstate123'; // YOUR gmail password
+                            // Sender and recipient settings
+                            $mail->setFrom('goldenstatecollege00@gmail.com', 'Goldentstate College');
+                            $mail->addAddress($email, $fname);
+                            $mail->addReplyTo('goldenstatecollege00@gmail.com', 'Goldentstate College'); // to set the reply to
+
+                            // Setting the email content
+                            $mail->IsHTML(true);
+                            $mail->Subject = "Goldenstate College Appointment System First Time Login Account";
+                            $mail->Body = 'Dear '.$fname.'<br><br>Your temporary username and password for <a href="gsck.online">Goldenstate College Appointment System</a> is ready to use.
+                                            <br><br>Username: '.$username.'<br>Password: '.$password.
+                                            '<br><br>If you have any concerns dont hesitate to contact us in goldenstatecollege00@gmail.com<br><br>Do not share your temporary username and password to anyone';
+                            $mail->AltBody = 'Dear '.$fname.'\n\nYour temporary account for Goldenstate College Appointment System(gsck.online) is ready to use.
+                            \n\nUsername: '.$username.'\nPassword: '.$password.
+                            '\n\nIf you have any concerns dont hesitate to contact us in goldenstatecollege00@gmail.com\n\nDo not share your temporary username and password to anyone';
+                            $mail->send();
+
+                    }
+
+                }
+                }
             }
         fclose($handle);
         ?>
@@ -103,7 +153,9 @@ else{
                             <input type="file" class="uploadfile" name="staff_file" accept=".csv, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet, application/vnd.ms-excel"/>
                             <input type="submit" class="btn_upload" name="upload" value="Upload" />
                         </div>
-
+                        <div>
+                        <input type="checkbox" name="check_list[]" value="true"><p>Overwrite Previous List</p>
+                        </div>
                         <p>
                             <?php
                                 $message = '';
@@ -152,7 +204,8 @@ else{
                     <p>Add a staff</p>
                     <input type="text"  id="lastname" name="lastname" placeholder="Last Name" required>
                     <input type="text"  id="firstname" name="firstname" placeholder="First Name" required>
-                    <input type="text"  id="staffid" name="staffid" placeholder="Staff ID" required> 
+                    <input type="text"  id="staffid" name="staffid" placeholder="Staff ID" required>
+                    <input type="text"  id="email" name="email" placeholder="Email Address" required>  
                     <div class="btn_add_container">
                         <button type="submit" class="btn_add" name="add">Add</button>
                     </div>
